# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AlertStatus": {
        "enum": [
          "new",
          "acknowledged",
          "investigating",
          "resolved",
          "false-positive"
        ],
        "title": "AlertStatus",
        "type": "string"
      },
...
      "AuditEvent": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "actor_role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OperatorRole"
              },
              {
                "type": "null"
              }
            ]
          },
          "actor_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor User Id"
          },
          "actor_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Username"
          },
          "details": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Details",
            "type": "object"
          },
          "event_id": {
            "title": "Event Id",
            "type": "integer"
          },
          "http_method": {
            "title": "Http Method",
            "type": "string"
          },
          "ip_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "resource_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Id"
          },
          "resource_type": {
            "title": "Resource Type",
            "type": "string"
          },
          "status_code": {
            "title": "Status Code",
            "type": "integer"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          }
        },
        "required": [
          "event_id",
          "occurred_at",
          "request_id",
          "action",
          "resource_type",
          "http_method",
          "path",
          "status_code"
        ],
        "title": "AuditEvent",
        "type": "object"
      },
...
      "CaseDisposition": {
        "enum": [
          "confirmed-fraud",
          "false-positive",
          "inconclusive",
          "referred-to-law-enforcement"
        ],
        "title": "CaseDisposition",
        "type": "string"
      },
      "CasePriority": {
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "CasePriority",
        "type": "string"
      },
      "CaseStatus": {
        "enum": [
          "open",
          "investigating",
          "escalated",
          "resolved",
          "closed"
        ],
        "title": "CaseStatus",
        "type": "string"
      },
...
      "CreateCaseFromAlertResult": {
        "additionalProperties": false,
        "properties": {
          "alert": {
            "$ref": "#/components/schemas/FraudAlert"
          },
          "case": {
            "$ref": "#/components/schemas/FraudCase"
          }
        },
        "required": [
          "alert",
          "case"
        ],
        "title": "CreateCaseFromAlertResult",
        "type": "object"
      },
...
      "FraudAlert": {
        "additionalProperties": false,
        "description": "An auto-generated or manually created alert that may lead to a case.",
        "properties": {
          "acknowledged_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged At"
          },
          "alert_id": {
            "title": "Alert Id",
            "type": "string"
          },
          "assigned_analyst_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Analyst Id"
          },
          "assigned_analyst_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Analyst Name"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "linked_case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linked Case Id"
          },
          "narrative": {
            "title": "Narrative",
            "type": "string"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "rule_code": {
            "title": "Rule Code",
            "type": "string"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "severity": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_type": {
            "$ref": "#/components/schemas/WorkflowSourceType",
            "default": "scenario"
          },
          "status": {
            "$ref": "#/components/schemas/AlertStatus",
            "default": "new"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "alert_id",
          "source_id",
          "rule_code",
          "title",
          "severity",
          "narrative",
          "created_at"
        ],
        "title": "FraudAlert",
        "type": "object"
      },
      "FraudCase": {
        "additionalProperties": false,
        "description": "A persistent fraud investigation case with lifecycle management.",
        "properties": {
          "alert_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Alert Count",
            "type": "integer"
          },
          "assigned_analyst_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Analyst Id"
          },
          "assigned_analyst_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Analyst Name"
          },
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "comment_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Comment Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "disposition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseDisposition"
              },
              {
                "type": "null"
              }
            ]
          },
          "priority": {
            "$ref": "#/components/schemas/CasePriority",
            "default": "medium"
          },
          "resolution_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Notes"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "risk_score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score",
            "type": "integer"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "sla_deadline": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sla Deadline"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_type": {
            "$ref": "#/components/schemas/WorkflowSourceType",
            "default": "scenario"
          },
          "status": {
            "$ref": "#/components/schemas/CaseStatus",
            "default": "open"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "case_id",
          "source_id",
          "title",
          "risk_score",
          "risk_level",
          "summary",
          "created_at",
          "updated_at"
        ],
        "title": "FraudCase",
        "type": "object"
      },
...
      "GetCurrentOperatorResult": {
        "additionalProperties": false,
        "properties": {
          "principal": {
            "$ref": "#/components/schemas/OperatorPrincipal"
          }
        },
        "required": [
          "principal"
        ],
        "title": "GetCurrentOperatorResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ListAlertsResult": {
        "additionalProperties": false,
        "properties": {
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/FraudAlert"
            },
            "title": "Alerts",
            "type": "array"
          },
          "page": {
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total_count": {
            "minimum": 0.0,
            "title": "Total Count",
            "type": "integer"
          }
        },
        "required": [
          "alerts",
          "total_count",
          "page",
          "page_size"
        ],
        "title": "ListAlertsResult",
        "type": "object"
      },
      "ListAuditEventsResult": {
        "additionalProperties": false,
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/AuditEvent"
            },
            "title": "Events",
            "type": "array"
          }
        },
        "required": [
          "events"
        ],
        "title": "ListAuditEventsResult",
        "type": "object"
      },
...
      "LoginCommand": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "maxLength": 256,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 64,
            "minLength": 3,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginCommand",
        "type": "object"
      },
      "LoginResult": {
        "additionalProperties": false,
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in_seconds": {
            "title": "Expires In Seconds",
            "type": "integer"
          },
          "principal": {
            "$ref": "#/components/schemas/OperatorPrincipal"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in_seconds",
          "principal"
        ],
        "title": "LoginResult",
        "type": "object"
      },
...
      "OperatorPrincipal": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "role": {
            "$ref": "#/components/schemas/OperatorRole"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "username",
          "display_name",
          "role",
          "is_active"
        ],
        "title": "OperatorPrincipal",
        "type": "object"
      },
      "OperatorRole": {
        "enum": [
          "analyst",
          "admin"
        ],
        "title": "OperatorRole",
        "type": "string"
      },
...
      "RiskLevel": {
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "RiskLevel",
        "type": "string"
      },
...
      "UpdateAlertStatusBody": {
        "additionalProperties": false,
        "properties": {
          "linked_case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linked Case Id"
          },
          "status": {
            "$ref": "#/components/schemas/AlertStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "UpdateAlertStatusBody",
        "type": "object"
      },
      "UpdateAlertStatusResult": {
        "additionalProperties": false,
        "properties": {
          "alert": {
            "$ref": "#/components/schemas/FraudAlert"
          }
        },
        "required": [
          "alert"
        ],
        "title": "UpdateAlertStatusResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkflowSourceType": {
        "enum": [
          "scenario",
          "dataset"
        ],
        "title": "WorkflowSourceType",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "A fraud triage workspace for uploaded transaction datasets, with persistent alerts and cases, plus reference scenario investigations for validation and rule calibration. Statistical and behavioral analysis remain the source of truth, while the optional copilot layer explains the results in operator-facing language.",
    "title": "Relational Fraud Intelligence",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/alerts": {
      "get": {
        "description": "Returns a paginated list of fraud alerts with optional status and severity filters.",
        "operationId": "list_alerts_api_v1_alerts_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AlertStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RiskLevel"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAlertsResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List fraud alerts",
        "tags": [
          "Alerts"
        ]
      }
    },
    "/api/v1/alerts/{alert_id}": {
      "patch": {
        "description": "Transitions an alert status and optionally links it to a case.",
        "operationId": "update_alert_status_api_v1_alerts__alert_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlertStatusBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateAlertStatusResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update alert status",
        "tags": [
          "Alerts"
        ]
      }
    },
    "/api/v1/alerts/{alert_id}/case": {
      "post": {
        "description": "Creates a persistent fraud case from an alert and links the alert to that case in a single workflow step.",
        "operationId": "create_case_from_alert_api_v1_alerts__alert_id__case_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCaseFromAlertResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create a case from an alert",
        "tags": [
          "Alerts"
        ]
      }
    },
    "/api/v1/audit-events": {
      "get": {
        "description": "Returns recent audit events. Admin only.",
        "operationId": "list_audit_events_api_v1_audit_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAuditEventsResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List audit events",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "get_current_operator_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCurrentOperatorResult"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get current operator profile",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/token": {
      "post": {
        "description": "Validates operator credentials and returns a JWT access token for subsequent API calls.",
        "operationId": "login_operator_api_v1_auth_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Authenticate an operator",
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
...
}